pyproject.toml
[tool.poetry]
name = "drf-plus-react-equals-love-backend-jonlee234"
version = "0.1.0"
description = ""
authors = ["Jonathane Leavell <jonathaneleavell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
